iphone 6/6 plus 対応

+++ MovieScore/MasterData.m
@@ -43,7 +43,7 @@ static MasterData* sharedHistory = nil;
 	return self;
 }
 
-- (unsigned long)retainCount
+- (NSUInteger)retainCount
 {
 	return UINT_MAX;
 }
